+++ src/main/kotlin/net/ndrei/teslapoweredthingies/machines/poweredkiln/PoweredKilnRecipes.kt
@@ -1,10 +1,5 @@
 package net.ndrei.teslapoweredthingies.machines.poweredkiln
 
-import net.minecraft.item.ItemStack
-import net.minecraft.item.crafting.FurnaceRecipes
-import net.ndrei.teslacorelib.utils.copyWithSize
-import net.ndrei.teslacorelib.utils.equalsIgnoreSize
-
 /**
  * Created by CF on 2017-07-06.
  */
@@ -17,8 +12,8 @@ object PoweredKilnRecipes {
 
     fun getRecipes() = this.recipes.toList()
 
-    fun findRecipe(input: ItemStack)
-            = this.recipes.firstOrNull { it.input.equalsIgnoreSize(input) } ?:
+    fun findRecipe(input: ItemStack) =
+        this.recipes.firstOrNull { it.input.equalsIgnoreSize(input) } ?:
             FurnaceRecipes.instance().getSmeltingResult(input).let {
                 if (it.isEmpty) null
                 else PoweredKilnRecipe(input.copyWithSize(1), it)